remove uneeded code
