+++ main.go
@@ -115,7 +115,7 @@ func main() {
 
 	cfg.Runtime = runtime
 
-	cfg.Runtime.QueryBuilder = query.NewMixedSourceQueryBuilder(
+	cfg.Runtime.QueryBuilder = query.NewSourceQueryBuilder(
 		cfg.Runtime.AURCache,
 		cfg.Runtime.Logger.Child("mixed.querybuilder"), cfg.SortBy,
 		cfg.Mode, cfg.SearchBy,
